requirements.txt
asgiref==3.2.3
blinker==1.4
certifi==2019.11.28
click==6.7
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
gunicorn==20.0.4
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mpmath==1.0.0
numpy==1.18.1
Pillow==7.0.0
# psycopg2==2.8.4
psycopg2-binary==2.8.3
pyparsing==2.4.6
pytz==2019.3
rdflib==4.2.2
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
sympy==1.1.1
Werkzeug==0.14.1
whitenoise==5.0.1
WTForms==2.1
